+++ packages/sdk/test/opRequestGas.test.ts
@@ -22,8 +22,8 @@ const DUMMY_GAS_ASSET: Asset = {
   id: ERC20_ID,
 };
 
-describe("prepareOperationRequest", async () => {
-  it("Has gas price 0 and dummy asset when opRequest gasPrice = 0", async () => {
+describe("handleGasForOperationRequest", async () => {
+  it("produces an operation request with gas price 0 and dummy gas assset when gasPrice set to 0", async () => {
     const [nocturneDB, merkleProver, signer, walletContract] = await setup(
       [500_000n, 500_000n],
       [shitcoin, shitcoin]
@@ -57,9 +57,9 @@ describe("prepareOperationRequest", async () => {
     expect(gasCompAccountedOpRequest.gasAsset).to.eql(DUMMY_GAS_ASSET);
   });
 
-  it("Accounts for gas asset when gasPrice > 0, same asset as joinsplit request", async () => {
+  it("adds gas comp to existing joinsplit when gas price is nonzero, ∃ a joinsplit unwrapping gasAsset, and user has enough", async () => {
     const [nocturneDB, merkleProver, signer, walletContract] = await setup(
-      [500_000n, 500_000n, 500_000n],
+      [500_000n, 500_000n, 2_000_000n],
       [shitcoin, shitcoin, shitcoin]
     );
     const deps = {
@@ -87,20 +87,18 @@ describe("prepareOperationRequest", async () => {
       opRequest
     );
 
-    const expectedJoinSplitRequest: JoinSplitRequest = {
-      asset: shitcoin,
-      unwrapValue: 1_350_000n, // 1M + 100k + 250k (250k is joinsplit + refund gas)
-    };
-
     expect(gasCompAccountedOpRequest.joinSplitRequests.length).to.eql(1);
-    expect(gasCompAccountedOpRequest.joinSplitRequests[0]).to.eql(
-      expectedJoinSplitRequest
-    );
     expect(gasCompAccountedOpRequest.gasPrice).to.eql(1n);
     expect(gasCompAccountedOpRequest.gasAsset).to.eql(shitcoin);
+
+    const gasAccountedJoinSplitRequest =
+      gasCompAccountedOpRequest.joinSplitRequests[0];
+    expect(gasAccountedJoinSplitRequest.asset).to.eql(shitcoin);
+    // exact amount depends
+    expect(gasAccountedJoinSplitRequest.unwrapValue > 1_000_000).to.be.true;
   });
 
-  it("Accounts for gas asset when gasPrice > 0, different joinsplit request for gas comp", async () => {
+  it("adds a joinsplit request including gas comp when gas price is nonzero, ∄ a joinsplit unwrapping a gasAsset, and user has enough of it", async () => {
     const [nocturneDB, merkleProver, signer, walletContract] = await setup(
       [500_000n, 500_000n, 2_000_000n],
       [shitcoin, shitcoin, stablescam]
@@ -134,19 +132,68 @@ describe("prepareOperationRequest", async () => {
       asset: shitcoin,
       unwrapValue: 1_000_000n, // just the 1M requested in unwrap
     };
-    const expectedJoinSplitRequestGas: JoinSplitRequest = {
-      asset: stablescam,
-      unwrapValue: 1_250_000n, // just the 1M requested in unwrap
-    };
+
+    expect(gasCompAccountedOpRequest.gasPrice).to.eql(1n);
+    expect(gasCompAccountedOpRequest.gasAsset).to.eql(stablescam);
 
     expect(gasCompAccountedOpRequest.joinSplitRequests.length).to.eql(2);
     expect(gasCompAccountedOpRequest.joinSplitRequests[0]).to.eql(
       expectedJoinSplitRequestUnwrap
     );
-    expect(gasCompAccountedOpRequest.joinSplitRequests[1]).to.eql(
-      expectedJoinSplitRequestGas
+
+    const joinSplitRequestForGas =
+      gasCompAccountedOpRequest.joinSplitRequests[1];
+    expect(joinSplitRequestForGas.asset).to.eql(stablescam);
+    // exact amount depends
+    expect(joinSplitRequestForGas.unwrapValue > 1_000_000).to.be.true;
+  });
+
+  it("adds a joinsplit request for gasAssetB when gas price is nonzeo, ∃ a joinsplit unwrapping gasAssetA, user doesn't have enough gasAssetA, but user does have enough gasAssetB", async () => {
+    const [nocturneDB, merkleProver, signer, walletContract] = await setup(
+      [500_000n, 500_000n, 500_000n, 2_000_000n],
+      [shitcoin, shitcoin, shitcoin, stablescam]
     );
-    expect(gasCompAccountedOpRequest.gasPrice).to.eql(1n);
-    expect(gasCompAccountedOpRequest.gasAsset).to.eql(stablescam);
+    const deps = {
+      walletContract,
+      merkle: merkleProver,
+      viewer: signer,
+      gasAssets: testGasAssets,
+      db: nocturneDB,
+    };
+
+    const builder = new OperationRequestBuilder();
+    const opRequest = builder
+      .action("0x1234", getDummyHex(0))
+      .unwrap(shitcoin, 100_000n)
+      .refundAsset(shitcoin)
+      .maxNumRefunds(1n)
+      .gas({
+        // Exceeds shitcoin balance, forces us to use stablescam
+        executionGasLimit: 1_000_000n,
+        gasPrice: 1n,
+      })
+      .build();
+
+    const gasCompAccountedOperationRequest = await handleGasForOperationRequest(
+      deps,
+      opRequest
+    );
+
+    expect(gasCompAccountedOperationRequest.gasPrice).to.eql(1n);
+    expect(gasCompAccountedOperationRequest.joinSplitRequests.length).to.equal(
+      2
+    );
+    expect(gasCompAccountedOperationRequest.gasAsset).to.eql(stablescam);
+
+    const joinSplitRequestForOp =
+      gasCompAccountedOperationRequest.joinSplitRequests[0];
+    expect(joinSplitRequestForOp.asset).to.eql(shitcoin);
+    expect(joinSplitRequestForOp.unwrapValue).to.eql(100_000n);
+
+    const joinSplitRequestForGas =
+      gasCompAccountedOperationRequest.joinSplitRequests[1];
+    expect(joinSplitRequestForGas.asset).to.eql(stablescam);
+    // exact amount depends
+    expect(joinSplitRequestForGas.unwrapValue >= 1_000_000).to.be.true;
   });
 });